requirements.txt
dlib==19.24.4
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
matplotlib==3.6.0


